[build] Refactor & centralize common Rust build script behavior.
This change extracts common code used in `build.rs` scripts to set
program versions for Habitat (userland) and Builder (hosted) components,
as well as generating protocol stubs.

While it is possible to use external crates in build scripts, it
involves adding a `build-dependencies` section in every Cargo.toml.
Instead, we're opting for a slightly "clever" of inlining Rust code at
compile time via the `include!()` macro. Currently, the two inlined
"libraies" are:

* `components/libbuild.rs` - contains common versioning logic for both
Builder and Habitat artifacts
* `components/libbuild-protocols.rs` - contains common logic to generate
Rust stubs for protocol files.

Additionally, the following refactorings we undertaken:

* Rename `components/bldr-build.rs` to `components/build-builder.rs` and
update call sites in related `Cargo.toml` files
* Rename `components/build.rs` to `components/builder-habitat.rs` and
update call sites in related `Cargo.toml` files

---

So...why?

While working on `builder-worker` I need to add more build-time
computing of dependent artifacts which led me to customize yet another
local `build.rs` file. After surveying the rest of the exceptional
components, it became clear that some consistency would help new
future components. Using the compile time inling trick, we get reuse
through composition when a local component needs specialization. Which
is needed in the immediate future work.
