+++ components/build-builder.rs
@@ -1,0 +1,6 @@
+// Inline common build behavior
+include!("libbuild.rs");
+
+fn main() {
+    builder::common();
+}
