+++ components/build-habitat.rs
@@ -1,0 +1,6 @@
+// Inline common build behavior
+include!("libbuild.rs");
+
+fn main() {
+    habitat::common();
+}
