--- components/build.rs
@@ -1,28 +1,0 @@
-use std::env;
-use std::fs::File;
-use std::io::BufReader;
-use std::io::prelude::*;
-use std::path::Path;
-
-fn main() {
-    let version = match env::var("PLAN_VERSION") {
-        Ok(ver) => ver,
-        _ => read_version(),
-    };
-    let mut f = File::create(Path::new(&env::var("OUT_DIR").unwrap()).join("VERSION")).unwrap();
-    f.write_all(version.trim().as_bytes()).unwrap();
-}
-
-fn read_version() -> String {
-    let ver_file = Path::new(&env::var("CARGO_MANIFEST_DIR").unwrap())
-        .parent()
-        .unwrap()
-        .parent()
-        .unwrap()
-        .join("VERSION");
-    let f = File::open(ver_file).unwrap();
-    let mut reader = BufReader::new(f);
-    let mut ver = String::new();
-    reader.read_line(&mut ver).unwrap();
-    ver
-}
