+++ components/builder-core/Cargo.toml
@@ -3,7 +3,6 @@ name = "builder_core"
 version = "0.0.0"
 authors = ["Adam Jacob <adam@chef.io>", "Jamie Winsor <reset@chef.io>", "Fletcher Nichol <fnichol@chef.io>", "Joshua Timberman <joshua@chef.io>", "Dave Parfitt <dparfitt@chef.io>", "Steven Murawski <smurawski@chef.io>"]
 workspace = "../../"
-build = "../bldr-build.rs"
 
 [dependencies]
 base64 = "*"
