+++ components/builder-depot/Cargo.toml
@@ -3,7 +3,6 @@ name = "habitat_depot"
 version = "0.0.0"
 authors = ["Adam Jacob <adam@chef.io>", "Jamie Winsor <reset@chef.io>", "Fletcher Nichol <fnichol@chef.io>", "Joshua Timberman <joshua@chef.io>", "Dave Parfitt <dparfitt@chef.io>"]
 description = "Habitat-Builder package depot"
-build = "../bldr-build.rs"
 workspace = "../../"
 
 [dependencies]
