+++ components/builder-graph/Cargo.toml
@@ -3,7 +3,7 @@ name = "habitat_builder_graph"
 version = "0.0.0"
 authors = ["The Habitat Maintainers <humans@habitat.sh>"]
 description = "Habitat-Builder Graph Dev Tool"
-build = "../bldr-build.rs"
+build = "../build-builder.rs"
 workspace = "../../"
 
 [[bin]]
