+++ components/builder-scheduler/Cargo.toml
@@ -3,7 +3,7 @@ name = "habitat_builder_scheduler"
 version = "0.0.0"
 authors = ["Adam Jacob <adam@chef.io>", "Jamie Winsor <reset@chef.io>", "Fletcher Nichol <fnichol@chef.io>", "Joshua Timberman <joshua@chef.io>", "Dave Parfitt <dparfitt@chef.io>"]
 description = "Habitat-Builder Job Scheduler"
-build = "../bldr-build.rs"
+build = "../build-builder.rs"
 workspace = "../../"
 
 [[bin]]
