+++ components/eventsrv-protocol/Cargo.toml
@@ -6,6 +6,7 @@ authors = [
   "Christopher Maier <cmaier@chef.io>",
   "Jamie Winsor <reset@habitat.sh>",
 ]
+build = "build.rs"
 workspace = "../../"
 
 [dependencies]
