+++ components/hab/Cargo.toml
@@ -2,7 +2,7 @@
 name = "hab"
 version = "0.0.0"
 authors = ["Adam Jacob <adam@chef.io>", "Jamie Winsor <reset@chef.io>", "Fletcher Nichol <fnichol@chef.io>", "Joshua Timberman <joshua@chef.io>", "Dave Parfitt <dparfitt@chef.io>"]
-build = "../build.rs"
+build = "../build-habitat.rs"
 workspace = "../../"
 
 [[bin]]
