+++ components/launcher-protocol/Cargo.toml
@@ -2,6 +2,7 @@
 name = "habitat-launcher-protocol"
 version = "0.0.0"
 authors = ["Jamie Winsor <reset@habitat.sh>"]
+build = "build.rs"
 
 [dependencies]
 protobuf = "*"
